BUGFIX: tracks with negative time left.
On some tracks the library com.google.android.exoplayer fails to detect the
end of the file and continuosly send time events even when it's not reproducing
anything. To fix this we control in the time events if the reproduction time is
longer that the one getted from the file and we skip to the next track.